+++ packages/core/src/hooks/useMode.ts
@@ -1,5 +1,5 @@
+import type { KubbFile } from '@kubb/fabric-core/types'
 import { useApp } from '@kubb/react-fabric'
-import type { KubbFile } from '../fs/index.ts'
 
 export function useMode(): KubbFile.Mode {
   const { meta } = useApp<{ mode: KubbFile.Mode }>()