+++ packages/plugin-oas/src/hooks/useOperationManager.ts
@@ -1,6 +1,6 @@
 import type { FileMetaBase, Plugin, ResolveNameParams } from '@kubb/core'
-import type { KubbFile } from '@kubb/core/fs'
 import { usePlugin, usePluginManager } from '@kubb/core/hooks'
+import type { KubbFile } from '@kubb/fabric-core/types'
 import type { Operation, Operation as OperationType } from '@kubb/oas'
 import { useContext } from '@kubb/react-fabric'
 import { Oas } from '../components/Oas.tsx'